backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
python-decouple
Pillow
django-ratelimit
whitenoise
gunicorn
psycopg2-binary
